# docs/swagger.yaml
basePath: /
definitions:
...
  dto.RouteCreateRequest:
    properties:
      allowed_lap_duration_seconds:
        type: integer
      checkpoint_geofence_id:
        type: string
      description:
        type: string
      geometry:
        type: string
      map_color:
        type: string
      max_speed_kmh:
        type: integer
      max_stop_duration_seconds:
        type: integer
      name:
        type: string
      stops:
        items:
          $ref: '#/definitions/dto.RouteStopRequest'
        type: array
    required:
    - name
    type: object
  dto.RouteFareCreateRequest:
    properties:
      amount_per_lap:
        type: number
      effective_from:
        type: string
      route_id:
        type: string
    required:
    - amount_per_lap
    - route_id
    type: object
  dto.RouteFareUpdateRequest:
    properties:
      amount_per_lap:
        type: number
      effective_from:
        type: string
      id:
        type: string
    type: object
  dto.RouteStopRequest:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      sequence:
        type: integer
    required:
    - name
    type: object
  dto.RouteUpdateRequest:
    properties:
      active:
        type: boolean
      allowed_lap_duration_seconds:
        type: integer
      checkpoint_geofence_id:
        type: string
      description:
        type: string
      geometry:
        type: string
      id:
        type: string
      map_color:
        type: string
      max_speed_kmh:
        type: integer
      max_stop_duration_seconds:
        type: integer
      name:
        type: string
      stops:
        items:
          $ref: '#/definitions/dto.RouteStopRequest'
        type: array
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RutaSegura API
  version: 0.1.2
paths:
...
  /api/route-fares:
    get:
      consumes:
      - application/json
      description: Get a list of all active route fares
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all route fares
      tags:
      - route-fares
    post:
      consumes:
      - application/json
      description: Set the amount charged per completed lap on a route, effective
        from a given date
      parameters:
      - description: Route Fare Create Request
        in: body
        name: route_fare
        required: true
        schema:
          $ref: '#/definitions/dto.RouteFareCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a new route fare
      tags:
      - route-fares
  /api/route-fares/{id}:
    get:
      consumes:
      - application/json
      description: Get a route fare by ID
      parameters:
      - description: Route Fare ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a route fare by ID
      tags:
      - route-fares
    put:
      consumes:
      - application/json
      description: Update the amount or effective date of an existing route fare
      parameters:
      - description: Route Fare ID
        in: path
        name: id
        required: true
        type: string
      - description: Route Fare Update Request
        in: body
        name: route_fare
        required: true
        schema:
          $ref: '#/definitions/dto.RouteFareUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an existing route fare
      tags:
      - route-fares
  /api/route-fares/{id}/status:
    patch:
      consumes:
      - application/json
      description: Toggle the status of a route fare
      parameters:
      - description: Route Fare ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Change status of a route fare (soft delete)
      tags:
      - route-fares
  /api/routes:
    get:
      consumes:
      - application/json
      description: Get a list of all active routes with their stops (public, no auth
        required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all routes
      tags:
      - routes
    post:
      consumes:
      - application/json
      description: Create a new bus route (line), optionally with its stops and GeoJSON
        geometry
      parameters:
      - description: Route Create Request
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/dto.RouteCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a new route
      tags:
      - routes
  /api/routes/{id}:
    get:
      consumes:
      - application/json
      description: Get a route with its stops and geometry by ID (public, no auth
        required)
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a route by ID
      tags:
      - routes
    put:
      consumes:
      - application/json
      description: Update an existing route; if stops is present, it fully replaces
        the existing stop list
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      - description: Route Update Request
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/dto.RouteUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an existing route
      tags:
      - routes
  /api/routes/{id}/eta:
    get:
      consumes:
      - application/json
      description: Get the estimated time of arrival of each live vehicle on a route
        to the queried lat/lon, sorted soonest-first (public, no auth required)
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      - description: Latitude of the query point
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude of the query point
        in: query
        name: lon
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Estimate arrival time of live vehicles to a point
      tags:
      - routes
...
  /api/routes/{id}/status:
    patch:
      consumes:
      - application/json
      description: Toggle the status of a route
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Change status of a route (soft delete)
      tags:
      - routes
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"